pyproject.toml
[tool.poetry]
name = "coinlist_helper"
version = "0.1.0"
description = ""
authors = ["Irishey <markirill21@gmail.com>"]

[tool.black]
line-length = 79


[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
alembic = "^1.7.6"
Flask-SQLAlchemy = "^2.5.1"
configparser = "^5.2.0"
Flask-Migrate = "^3.1.0"
Flask-Script = "^2.0.6"
psycopg2 = "^2.9.3"
python-dotenv = "^0.19.2"
Flask-RESTful = "^0.3.9"
Flask-WTF = "^1.0.0"
Flask-Login = "^0.5.0"
Flask-RBAC = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
